requirements.txt
Flask==3.1.2
psycopg2-binary==2.9.11
Werkzeug==3.1.3
requests==2.32.0
python-dotenv==1.0.0
